requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pandas==2.2.0
pillow==10.2.0
plotly==5.19.0
scikit-learn==1.4.0
seaborn==0.13.2
streamlit==1.30.0
